B-26 added minimal settings + refactoring
