+++ .gitignore
@@ -1,9 +1,9 @@
 HELP.md
 target/
+!.mvn/wrapper/maven-wrapper.jar
 !**/src/main/**
 !**/src/test/**
-.idea/
-.mvn/
+
 ### STS ###
 .apt_generated
 .classpath
@@ -16,8 +16,8 @@ target/
 ### IntelliJ IDEA ###
 .idea
 *.iws
-*.ipr
 *.iml
+*.ipr
 
 ### NetBeans ###
 /nbproject/private/
