--- endpoint/src/main/java/platformbackend/endpoint/EndpointApplication.java
@@ -1,13 +1,0 @@
-package platformbackend.endpoint;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EndpointApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(EndpointApplication.class, args);
-    }
-
-}
