--- endpoint/src/main/java/platformbackend/endpoint/rest/MainController.java
@@ -1,17 +1,0 @@
-package platformbackend.endpoint.rest;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-
-@RestController
-@RequestMapping("/")
-public class MainController {
-
-    @GetMapping
-    public ResponseEntity getMain() {
-        return ResponseEntity.ok("Hello world");
-    }
-}
