+++ endpoint/src/main/java/ru/studentsplatform/backend/endpoint/EndpointApplication.java
@@ -1,0 +1,17 @@
+package ru.studentsplatform.backend.endpoint;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.ComponentScan;
+
+@SpringBootApplication
+@ComponentScan(basePackages = {"ru.studentsplatform.backend.*"})
+@EntityScan(basePackages = {"ru.studentsplatform.backend.*"})
+public class EndpointApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(EndpointApplication.class, args);
+	}
+
+}
