--- endpoint/src/main/resources/templates/main.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>MainPage</title>
-</head>
-<body>
-Hello world!
-</body>
-</html>
