+++ entities/src/main/java/ru/studentsplatform/backend/entities/model/UnuniversityEnum.java
@@ -1,0 +1,11 @@
+package ru.studentsplatform.backend.entities.model;
+
+public enum  UnuniversityEnum {
+	SPBU("Санкт-Петербургский государственный университет");
+
+	private String name;
+
+	UnuniversityEnum(String name) {
+		this.name = name;
+	}
+}
