--- entities/src/test/java/entities/entities/package-info.java
@@ -1,1 +1,0 @@
-package entities.entities;
