+++ pom.xml
@@ -21,6 +21,7 @@
         <java.version>11</java.version>
         <lombok.version>1.18.12</lombok.version>
         <maven-checkstyle-plugin.version>3.1.1</maven-checkstyle-plugin.version>
+        <micrometer-registry-prometheus.version>1.5.2</micrometer-registry-prometheus.version>
     </properties>
 
     <modules>
@@ -53,7 +54,11 @@
                 <type>pom</type>
             </dependency>
 
-
+            <dependency>
+                <groupId>io.micrometer</groupId>
+                <artifactId>micrometer-registry-prometheus</artifactId>
+                <version>${micrometer-registry-prometheus.version}</version>
+            </dependency>
         </dependencies>
     </dependencyManagement>
 
@@ -77,7 +82,7 @@
                     </dependency>
                 </dependencies>
                 <configuration>
-                    <configLocation>/config/google_checks.xml</configLocation>
+                    <configLocation>/config/codestyle_check_config.xml</configLocation>
                     <encoding>UTF-8</encoding>
                     <consoleOutput>true</consoleOutput>
                     <failsOnError>true</failsOnError>
@@ -114,7 +119,7 @@
                 <artifactId>maven-checkstyle-plugin</artifactId>
                 <version>3.1.1</version>
                 <configuration>
-                    <configLocation>config/google_checks.xml</configLocation>
+                    <configLocation>config/codestyle_check_config.xml</configLocation>
                     <headerLocation>/config/LICENSE.txt</headerLocation>
                 </configuration>
             </plugin>
