--- system/src/main/java/platformbackend/system/utils/IpReceiver.java
@@ -1,9 +1,0 @@
-package platformbackend.system.utils;
-
-import javax.servlet.http.HttpServletRequest;
-
-public class IpReceiver {
-    public static String getIp(HttpServletRequest httpServletRequest) {
-        return httpServletRequest.getRemoteAddr();
-    }
-}
