+++ telegram-bot/src/main/java/ru/studentsplatform/backend/tlgrmbot/config/TelegramBotConfig.java
@@ -1,0 +1,7 @@
+package ru.studentsplatform.backend.tlgrmbot.config;
+
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class TelegramBotConfig {
+}
