+++ university/src/main/java/ru/studentsplatform/backend/university/util/UniversityUtilTest.java
@@ -1,0 +1,7 @@
+package ru.studentsplatform.backend.university.util;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class UniversityUtilTest {
+}
