--- university/src/main/java/university/package-info.java
@@ -1,1 +1,0 @@
-package university;
