+++ vk-bot/src/main/java/ru/studentsplatform/backend/vkbot/util/VkBotUtilTest.java
@@ -1,0 +1,7 @@
+package ru.studentsplatform.backend.vkbot.util;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class VkBotUtilTest {
+}
