--- vk-bot/src/main/java/vkbot/package-info.java
@@ -1,1 +1,0 @@
-package vkbot;
