--- vk-bot/src/test/java/vkbot/package-info.java
@@ -1,1 +1,0 @@
-package vkbot;
